Add partial support for referencing a target more than once
If you tweened the same target more than once inside another movie
clip, whether in the same tween or in another tween, the parser would
error out. This tracks when a target has already been processed so
that the second time it can still follow the temporary to final rename
chaining.

Not sure that this is very useful, given that the rest of it may not
really support multiple tweens on the same target and animation will
get funky, but at least it doesn't crash.